// src/types/user.ts
export type User = {
  image: string;
  username: string;
  email: string;
  category: string;
  sold: number;
  profit: number;
  location: string;
  aboutme: string;
  _id: string;
  phone: string;
  fullname: string;
  jobRole: string;
  experience: string;
  address: string;
  profileImage: string;
  phoneNumber: string;  // Change to string
  notifications: string[];  // Change to array of strings
};